refactor: move map search to services directory
